requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
dj-database-url==2.2.0
Django==5.1
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==2.2.0
django-on-heroku==1.1.2
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.2
setuptools==74.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
